infrastructure/requirements.txt
aws_cdk.core==1.91.0
chalice[cdk]==1.22.1
